requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-crispy-forms==1.6.1
django-extensions==1.7.6
python-decouple
uWSGI==2.0.14
psycopg2==2.6.1
django-bootstrap3
boto
django-storages
django-test-without-migrations==0.4
factory_boy